Some UI utilities was incorrectly placed in the util.py file. They have been moved to ui_util.py file in stead.
