+++ ui_widgets_old.py
@@ -3,8 +3,7 @@ from PyQt5.QtCore import Qt, QDir, QDateTime, QDate, QThread,QObject,QModelIndex
 from PyQt5.QtGui import QPixmap
 import settings
 from file_metadata_util import FileMetadata, StandardizeFilenames, CopyLogicalTags, ConsolidateMetadata
-from util import clearLayout, AutoCompleteList
-from ui_util import ProgressBarWidget
+from ui_util import ProgressBarWidget, AutoCompleteList, clearLayout
 import os
 from file_util import FileRenamer
